refactor : use Biome and Prettier only in the root and move the types to `@yadwy/types` pkg
