--- apps/store/.prettierignore
@@ -1,13 +1,0 @@
-# handled by Biome
-*.js
-*.ts
-*.cjs
-*.mjs
-*.d.cts
-*.d.mts
-*.jsx
-*.tsx
-*.json
-*.jsonc
-*.css
-*.scss
