+++ apps/store/src/lib/product-utils.ts
@@ -1,4 +1,4 @@
-import type { Product } from "~/types";
+import type { Product } from "@yadwy/types";
 
 export function getProductCoverImage(product: Pick<Product, "images">) {
   return product.images[0] || "/placeholder.svg";
