+++ package.json
@@ -7,11 +7,17 @@
     "dev": "bun store dev",
     "store": "turbo --filter @yadwy/store",
     "prepare": "lefthook install",
-    "check": "turbo check",
+    "check": "bun check:biome && bun check:prettier && bun check:types",
     "check:types": "turbo check:types",
+    "check:prettier": "prettier . --write",
+    "check:biome": "biome check . --write --unsafe",
     "check:biome:ci": "biome ci . --formatter-enabled=true --linter-enabled=true"
   },
   "devDependencies": {
-    "@yadwy/config": "workspace:*"
+    "@yadwy/config": "workspace:*",
+    "@biomejs/biome": "^1.9.4",
+    "lefthook": "^1.11.10",
+    "prettier": "^3.5.3",
+    "turbo": "^2.5.0"
   }
 }
