+++ packages/types/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@yadwy/types",
+  "private": true,
+  "main": "src/index.ts",
+  "scripts": {
+    "check:types": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "tailwindcss": "^4.1.4"
+  }
+}
