+++ packages/types/src/entity.ts
@@ -1,0 +1,3 @@
+export interface Entity {
+  id: number;
+}
