+++ packages/types/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@yadwy/config/tsconfig.json",
+  "include": ["**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "exclude": ["node_modules", ".next", ".turbo"]
+}
